Move elastic_filter to elastic/filter
