+++ elastic/insta_user-elastic_filter/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"strings"
 
-	"github.com/codeuniversity/smag-mvp/kafka/changestream"
+	"github.com/codeuniversity/smag-mvp/elastic/filter"
 	"github.com/codeuniversity/smag-mvp/service"
 	"github.com/codeuniversity/smag-mvp/utils"
 )
@@ -16,7 +16,7 @@ func main() {
 	elasticAddresses := utils.GetStringFromEnvWithDefault("ELASTIC_ADDRESSES", "http://localhost:9200,http://localhost:9201")
 	splitAddr := strings.Split(elasticAddresses, ",")
 
-	f := changestream.NewKafkaToElasticFilter(kafkaAddress, groupID, changesTopic, splitAddr)
+	f := filter.NewKafkaToElasticFilter(kafkaAddress, groupID, changesTopic, splitAddr)
 
 	service.CloseOnSignal(f)
 	waitUntilClosed := f.Start()
